Former-commit-id: 0556726833def25002fc91d8e2141c8f9ee47955
Former-commit-id: eed3ab646613bdc5a2488c733292788879da423f

+++ src/simulation/architecture/messaging/system_messaging.cpp
@@ -88,7 +88,7 @@ void SystemMessaging::selectMessageBuffer(int64_t bufferUse)
     std::vector<MessageStorageContainer*>::iterator it;
     it = this->dataBuffers.begin();
 
-    if(bufferUse >= this->dataBuffers.size())
+    if(bufferUse >= (int64_t) this->dataBuffers.size())
     {
         BSK_PRINT(MSG_ERROR,"You've attempted to access a message buffer that does not exist. Yikes.");
         this->messageStorage = *this->dataBuffers.begin();
